:recycle: Refactor: 비밀번호 변경 페이지- 함수 분리, 파일경로 alias로 수정함
